+++ src/mock_mmu.h
@@ -12,7 +12,7 @@ public:
   void write(uint16_t addr, uint16_t word) override;
 
   [[nodiscard]] uint8_t read8(uint16_t addr) const override;
-  [[nodiscard]] uint8_t read16(uint16_t addr) const override;
+  [[nodiscard]] uint16_t read16(uint16_t addr) const override;
 
   void inc(uint16_t addr) override;
